convert terrain to double precision

+++ terrain_util.h
@@ -36,8 +36,8 @@ void  terrain_coord2xy(double lat, double lon,
 void  terrain_xy2coord(float x, float y,
                        double* lat, double* lon);
 void  terrain_geo2xyz(double lat, double lon, float alt,
-                      float* x, float* y, float* z);
-void  terrain_xyz2geo(float x, float y, float z,
+                      double* x, double* y, double* z);
+void  terrain_xyz2geo(double x, double y, double z,
                       double* lat, double* lon, float* alt);
 void  terrain_xyz2xyh(float x1, float y1, float z1,
                       float* x2, float* y2, float* alt);
